feat: Structurizr support to Aspire.Hosting
This commit includes the introduction of Structurizr support to the .NET Aspire project. New members, methods and classes relevant to the Structurizr resource were added, along with their respective licenses. Further changes include the creation of StructurizrResource.cs, which aids in the configuration of a Structurizr server, and modifications to the Aspire.Hosting.Structurizr.csproj file. Furthermore, a README.md file was added to provide instructions on how to integrate and use the Structurizr support in the Aspire project.