docs: Update Aspire.Hosting.Structurizr README reference
The reference to the README.md file within the Aspire.Hosting.Structurizr project has been updated. The previous README file was deleted and references in the project file were updated to point to the README file from the parent directory. This change ensures the right README file gets bundled with the package.